resume-generator/pyproject.toml

[project]
name = "radium226-resume-generator"
version = "0.1.0"
description = "Radium226's Resume Generator"
authors = [
    {name = "Radium226", email = "radium226@rouages.xyz"},
]
dependencies = [
    "click>=8.1.3",
    "pydantic>=1.10.9",
    "pydantic-yaml>=0.11.2",
    "ruamel-yaml>=0.17.31",
    "mistletoe>=1.1.0",
    "pendulum>=2.1.2",
    "lxml>=4.9.2",
]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "GPL"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[project.scripts]
radium226-resume-generator = "radium226.resume_generator:app"

[tool.pdm.build]
package-dir = "src" 
includes = ["src/radium226/resume_generator"]
[tool.pdm.dev-dependencies]
dev = [
    "pytest>=7.3.2",
    "mypy>=1.3.0",
]
